requirements.txt
bokeh
boto3
bottleneck
dask
distributed
folium
jinja2
matplotlib
odc-geo
odc-stac
pystac
pystac_client
rio-stac
scipy
